implement first fetch
Due to API rate limiting, I get api response as json and stored it in
data.json file for testing.
Here I'm using Intersection observer API to check if fetch box is
visible in the viewport; If so we need to fetch images from API (in this
case I'm fetching from data.json file).
When we get the data, we clone the image component and insert all the
data into it; But more importantly, we insert image url to data-src
instead of src attribute, and that will help us to implement lazy
loading feature which will be implemented next.